+++ src/models/mod.rs
@@ -1,3 +1,4 @@
 pub mod events;
+pub mod plate_appearance;
 pub mod play_ball;
 pub mod types;